requirements.txt
gdown==3.13.0
pytorch_lightning==1.3.1
torch==1.8.1
torchvision==0.9.1
Pillow==8.2.0
